Resolve puppet overwriting a pre-existing bootstrap-osd auth entry
Ceph jewel creates the client.bootstrap-osd auth entry when the cluster
is created. puppet-ceph has been overwriting this keyring in memory and
on the filesystem. This patch updates ceph::osd to exract the
client keyring from memory and write it to the filesystem. This will
means ceph::key will not need to create the client.bootstrap-osd keyring.

Change-Id: I11f134bed59da706a03ce4a70d941145c16f8175
Partial-bug: 1769722